leader election -> dedicated pod
