+++ src/cli/dev.ts
@@ -5,7 +5,7 @@ import { ChildProcess, fork } from "child_process";
 import { promises as fs } from "fs";
 import prompt from "prompts";
 
-import { Webhook } from "../lib/k8s/webhook";
+import { Assets } from "../lib/assets";
 import Log from "../lib/logger";
 import { buildModule, loadModule } from "./build";
 import { RootCmd } from "./root";
@@ -35,7 +35,7 @@ export default function (program: RootCmd) {
       const { cfg, path } = await loadModule();
 
       // Generate a secret for the module
-      const webhook = new Webhook(
+      const webhook = new Assets(
         {
           ...cfg.pepr,
           description: cfg.description,
