+++ src/lib/assets/webhooks.ts
@@ -1,0 +1,186 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: 2023-Present The Pepr Authors
+
+import {
+  AdmissionregistrationV1WebhookClientConfig as AdmissionRegnV1WebhookClientCfg,
+  V1LabelSelectorRequirement,
+  V1MutatingWebhookConfiguration,
+  V1RuleWithOperations,
+  V1ValidatingWebhookConfiguration,
+} from "@kubernetes/client-node";
+import { fork } from "child_process";
+import { concat, equals, uniqWith } from "ramda";
+
+import { Assets } from ".";
+import Log from "../logger";
+import { Binding, Event } from "../types";
+
+const peprIgnoreLabel: V1LabelSelectorRequirement = {
+  key: "pepr.dev",
+  operator: "NotIn",
+  values: ["ignore"],
+};
+
+const peprIgnoreNamespaces: string[] = ["kube-system", "pepr-system"];
+
+export function generateWebhookRules(
+  assets: Assets,
+  isMutateWebhook: boolean,
+  path: string,
+): Promise<V1RuleWithOperations[]> {
+  return new Promise((resolve, reject) => {
+    const rules: V1RuleWithOperations[] = [];
+
+    // Fork is needed with the PEPR_MODE env var to ensure the module is loaded in build mode and will send back the capabilities
+    const program = fork(path, {
+      env: {
+        ...process.env,
+        LOG_LEVEL: "warn",
+        PEPR_MODE: "build",
+      },
+    });
+
+    // We are receiving javascript so the private fields are now public
+    interface ModuleCapabilities {
+      _name: string;
+      _description: string;
+      _namespaces: string[];
+      _bindings: Binding[];
+    }
+
+    // Wait for the module to send back the capabilities
+    program.on("message", message => {
+      // Cast the message to the ModuleCapabilities type
+      const capabilities = message.valueOf() as ModuleCapabilities[];
+
+      // Iterate through the capabilities and generate the rules
+      for (const capability of capabilities) {
+        Log.info(`Module ${assets.config.uuid} has capability: ${capability._name}`);
+
+        // Read the bindings and generate the rules
+        for (const binding of capability._bindings) {
+          const { event, kind, isMutate, isValidate } = binding;
+
+          // If the module doesn't have a callback for the event, skip it
+          if (isMutateWebhook && !isMutate) {
+            continue;
+          }
+
+          if (!isMutateWebhook && !isValidate) {
+            continue;
+          }
+
+          const operations: string[] = [];
+
+          // CreateOrUpdate is a Pepr-specific event that is translated to Create and Update
+          if (event === Event.CreateOrUpdate) {
+            operations.push(Event.Create, Event.Update);
+          } else {
+            operations.push(event);
+          }
+
+          // Use the plural property if it exists, otherwise use lowercase kind + s
+          const resource = kind.plural || `${kind.kind.toLowerCase()}s`;
+
+          rules.push({
+            apiGroups: [kind.group],
+            apiVersions: [kind.version || "*"],
+            operations,
+            resources: [resource],
+          });
+        }
+      }
+    });
+
+    program.on("exit", code => {
+      if (code !== 0) {
+        reject(new Error(`Child process exited with code ${code}`));
+      } else {
+        // If there are no rules, add a catch-all
+        if (rules.length < 1) {
+          resolve(rules);
+        } else {
+          const reducedRules = uniqWith(equals, rules);
+          resolve(reducedRules);
+        }
+      }
+    });
+
+    program.on("error", error => {
+      reject(error);
+    });
+  });
+}
+
+export async function webhookConfig(
+  assets: Assets,
+  mutateOrValidate: "mutate" | "validate",
+  path: string,
+  timeoutSeconds = 10,
+): Promise<V1MutatingWebhookConfiguration | V1ValidatingWebhookConfiguration | null> {
+  const ignore = [peprIgnoreLabel];
+
+  const { name, tls, config, apiToken, host } = assets;
+  const ignoreNS = concat(peprIgnoreNamespaces, config.alwaysIgnore.namespaces || []);
+
+  // Add any namespaces to ignore
+  if (ignoreNS) {
+    ignore.push({
+      key: "kubernetes.io/metadata.name",
+      operator: "NotIn",
+      values: ignoreNS,
+    });
+  }
+
+  const clientConfig: AdmissionRegnV1WebhookClientCfg = {
+    caBundle: tls.ca,
+  };
+
+  // The URL must include the API Token
+  const apiPath = `/${mutateOrValidate}/${apiToken}`;
+
+  // If a host is specified, use that with a port of 3000
+  if (host) {
+    clientConfig.url = `https://${host}:3000${apiPath}`;
+  } else {
+    // Otherwise, use the service
+    clientConfig.service = {
+      name: name,
+      namespace: "pepr-system",
+      path: apiPath,
+    };
+  }
+
+  const isMutate = mutateOrValidate === "mutate";
+  const rules = await generateWebhookRules(assets, isMutate, path);
+
+  // If there are no rules, return null
+  if (rules.length < 1) {
+    return null;
+  }
+
+  return {
+    apiVersion: "admissionregistration.k8s.io/v1",
+    kind: isMutate ? "MutatingWebhookConfiguration" : "ValidatingWebhookConfiguration",
+    metadata: { name },
+    webhooks: [
+      {
+        name: `${name}.pepr.dev`,
+        admissionReviewVersions: ["v1", "v1beta1"],
+        clientConfig,
+        failurePolicy: "Ignore",
+        matchPolicy: "Equivalent",
+        timeoutSeconds,
+        namespaceSelector: {
+          matchExpressions: ignore,
+        },
+        objectSelector: {
+          matchExpressions: ignore,
+        },
+        rules,
+        // @todo: track side effects state
+        sideEffects: "None",
+      },
+    ],
+  };
+}
